Add executionHash transmission to l1; Add tests
